Separate code into compoennts for strength/weakness cards
